fix phase 4
